+++ core/templates/services/request-interceptor.service.ts
@@ -22,42 +22,6 @@ import { Injectable } from '@angular/core';
 import { switchMap } from 'rxjs/operators';
 import { CsrfTokenService } from './csrf-token.service';
 
-@Injectable({
-  providedIn: 'root'
-})
-export class MockCsrfTokenService {
-  // 'tokenPromise' will be null when token is not initialized .
-  tokenPromise: PromiseLike<string> | null = null;
-
-  initializeToken(): void {
-    if (this.tokenPromise !== null) {
-      throw new Error('Token request has already been made');
-    }
-    // TODO(#8035): Remove the use of $.ajax and hence the ts-ignore
-    // in csrf-token.service.spec.ts once all the services are migrated
-    // We use jQuery here instead of Angular's $http, since the latter creates
-    // a circular dependency.
-    this.tokenPromise = $.ajax({
-      url: '/csrfhandler',
-      type: 'GET',
-      dataType: 'text',
-      dataFilter: function(data: string) {
-        // Remove the protective XSSI (cross-site scripting inclusion) prefix.
-        let actualData = data.substring(5);
-        return JSON.parse(actualData);
-      },
-    }).then(function(response: {token: string}) {
-      return response.token;
-    });
-  }
-
-  getTokenAsync(): PromiseLike<string> {
-    if (this.tokenPromise === null) {
-      throw new Error('Token needs to be initialized');
-    }
-    return this.tokenPromise;
-  }
-}
 
 @Injectable({
   providedIn: 'root'
